Rework the generation of environment for the application.
This ensure that only the explicit set variables in s-salome-bin are
generated and also it makes sure all of them are generated. For this
we now use a sentinel variable.

Also, we fix the generated environment for s-salome-tar as the final
directories for SALOME_BIN_APPLICATION and SALOME_BIN_MODULES must
be pointing into the installed binrary release.

Last also allow for more variables to be generated for the Python
part. None where generated for s-salome-bin for example.

Should fix #26613.