s-salome-tar: Do not use KERNEL_ROOT_DIR to point to modules.
The CMake files from the binary tar have some hard coded path coming
from Jenkins (/volatile/...). To fix that the environment variable
KERNEL_ROOT_DIR was used as reference to point to modules in the
binary archive. This is wrong as soon as the KERNEL module is
built in NG mode.

Now the reference to the modules in the binary archive is using
SALOME_BIN_MODULES from s-salome-tar env.

This fix KERNEL build in NG mode.