s-cmake/s-swig: Do not set PATH for native variant.
The PATH should already be set anyway and if the native version is
needed the plug-ing must not make another one available.

This should help clean-up some build issue when a plug-in adding
a standard path like /usr/bin in front of the PATH variable makes
visible another tool that was compiled in the sandbox and now
not the one that will be used.