Remove installed MeshGems to avoid conflicts with new one.
This is needed on Windows as the MeshGems install is done in a shared
directory and so even after commenting out all environment variables
for the MeshGems module itself there is still a PATH pointing to the
MeshGems binaries.

Note that with this patch we also clean the MeshGems install on
GNU/Linux.