Fix setup script infinite loop.
This is a trial/error kind of fix. The origin of the issue was a
change from:
  while [ ! -z $1 ]; do
to
  while [ -n $1 ]; do
Which should be strictly equivalent. But for some unknown reason
it is not and the later enter an infinite loop. After some testing
it turns out that using Bash built-in [[ ]] fixes this issue.